demo/analytics/requirements.txt
Flask==2.3.2
matplotlib==3.7.1
pandas==2.0.2
plotly==5.15.0
python-dotenv==1.0.0
kaleido==0.2.1
